third_party/blink/web_tests/external/wpt/app-history/navigate-event/navigate-appHistory-navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_true(e.cancelable);
    assert_true(e.canRespond);
    assert_false(e.userInitiated);
    assert_true(e.hashChange);
    assert_equals(new URL(e.destination.url).hash, "#foo");
    assert_true(e.destination.sameDocument);
    assert_equals(e.formData, null);
  });
  appHistory.navigate("#foo");
}, "navigate event for appHistory.navigate()");
</script>
